+++ application/shared/view/timeclock/jobs/print.php
@@ -36,30 +36,66 @@ switch ($this->sys->template->job['status']) {
         </tr>
     </tfoot>
 </table>
-<h2>Hours Breakdown ({total_hours} Total Hours)</h2>
+<h2>Hours Breakdown ({total_hours} Total Hours | <?php echo $this->model_jobs->quoted_hours($this->sys->template->job['job_id']); ?> Quoted Hours)</h2>
 <div class="row">
     <div class="col-sm-6">
         <table class="table">
             <thead>
                 <tr>
                     <th>Category</th>
-                    <th>Total Hours</th>
+                    <th>Worked Load</th>
+                    <th>Quoted Load</th>
+                    <th>Worked Hours</th>
+                    <th>Quoted Hours</th>
+                    <th>Difference</th>
                 </tr>
             </thead>
             <tbody>
                 <?php
-                foreach ($this->sys->template->hours_by_category as $category=>$time) {
-                    echo '<tr>';
-                    echo '<td>' . $category . '</td>';
-                    echo '<td>' . $time . '</td>';
-                    echo '</tr>';
+                foreach ($this->sys->template->job['quoted_time'] as $category=>$time) {
+                    if (0 < $time) {
+                        echo '<tr>';
+                            echo '<td>' . $this->sys->template->categories[$category]['category_name'] . '</td>';
+                            
+                            if (array_key_exists($category, $this->sys->template->worked_load)) {
+                                echo '<td>' . $this->sys->template->worked_load[$category] . '%</td>';
+                            } else {
+                                echo '<td>0%</td>';
+                            }
+                            if (array_key_exists($category, $this->sys->template->quoted_load)) {
+                                echo '<td>' . $this->sys->template->quoted_load[$category] . '%</td>';
+                            } else {
+                                echo '<td>0%</td>';
+                            }
+                            
+                            if (array_key_exists($category, $this->sys->template->hours_by_category)) {
+                                $worked = $this->sys->template->hours_by_category[$category];
+                                echo '<td>' . $this->sys->template->hours_by_category[$category] . '</td>';
+                            } else {
+                                $worked = 0;
+                                echo '<td>0</td>';
+                            }
+                            echo '<td>' . $time . '</td>';
+                            
+                            $difference = $time - $worked;
+                            if (0 <= $difference) {
+                                echo '<td class="green">' . $difference . '</td>';
+                            } else {
+                                echo '<td class="red">' . $difference . '</td>';
+                            }
+                        echo '</tr>';
+                    }
                 }
                 ?>
             </tbody>
             <tfoot>
                 <tr>
                     <th>Category</th>
-                    <th>Total Hours</th>
+                    <th>Worked Load</th>
+                    <th>Quoted Load</th>
+                    <th>Worked Hours</th>
+                    <th>Quoted Hours</th>
+                    <th>Difference</th>
                 </tr>
             </tfoot>
         </table>